chore: Update Tooltip component to use arrowColor middleware
